+++ src/ui/map.js
@@ -1,5 +1,5 @@
 import {Component, html} from './lib.js'
-import {random as randomBetween} from '../game/utils.js'
+import {random as randomBetween} from '../utils.js'
 import {isRoomCompleted} from '../game/utils-state.js'
 import {roomTypes} from '../game/dungeon-rooms.js'
